pos.Info: Merge into goldast
With the removal of the pos.Pos type,
and moving ErrorList into the goldast package,
the only use of the pos package is for the Info type,
which is used exclusively by goldsat.Parse.

Merge it all into one package.
